bk file-loader isolated; k-v things isolated
