+++ src/load/bkidx.cpp
@@ -3,54 +3,17 @@
  */
 #include <iostream>
 #include <filesystem>
+//#include <cstdint>
+//#include <fstream>
 
-#include "bce.h"
 #include "bkidx.h"
-#include <queue>
+#include "datfarm.h"
 
 using namespace std;
 
-FOFF_T   *FOFF;
+FOFF_T    *FOFF;
+DATFARM_T *datfarm;
 const uint32_t  BK_SIGN = 0xD9B4BEF9;   // LE
-const uint8_t WIN_SIZE = 16;
-static queue<size_t> opened;
-
-bool        DATFARM_T::open(const size_t no)
-{
-    if (no > qty) {
-        cerr << "File # too big: " << no << endl;
-        return false;
-    }
-    if (file[no].is_open())
-        return true;
-    if (opened.size() >= WIN_SIZE) {  // close unused file
-        file[opened.front()].close();
-        opened.pop();
-    }
-    ostringstream ss;
-    ss << setw(5) << setfill('0') << no;
-    string fn = folder + "blk" + ss.str() + ".dat";
-    file[no].open(fn, ios::in|ios::binary);
-    if (!file[no].is_open()) {
-        cerr << "Can't open file " << fn << endl;
-        return false;
-    }
-    opened.push(no);
-    return true;
-}
-
-bool        DATFARM_T::read(const size_t no, const size_t offset, const int size, void *dst)
-{
-    if (!open(no))
-        return false;
-    file[no].seekg(offset, file[no].beg);
-    file[no].read(static_cast<char *>(dst), size);
-    if (file[no].gcount() != size) {
-        cerr << "Can't read " << size << " bytes from " << no << "." << offset << endl;
-        return false;
-    }
-    return true;
-}
 
 size_t  load_fileoffsets(const string fn)  ///< load file-offset file
 {
@@ -78,10 +41,21 @@ size_t  load_fileoffsets(const string fn)  ///< load file-offset file
     return blocks;
 }
 
-bool    load_bk(DATFARM_T &datfarm, const uint32_t fileno, const uint32_t offset)       ///< load bk to buffer
+size_t  init_bkloader(const std::string datdir, const std::string locsfile)
 {
+  auto bk_qty = load_fileoffsets(locsfile);
+  if (bk_qty)
+    datfarm = new DATFARM_T(bk_qty, (datdir.back() == '/') ? datdir : datdir + '/');
+  return bk_qty;
+}
+
+bool    load_bk(const uint32_t bk_no, char *dst)       ///< load bk to buffer
+{
+  // load_bk(datfarm, FOFF[COUNT.bk].fileno, FOFF[COUNT.bk].offset)
+  // load_bk(DATFARM_T &datfarm, const uint32_t fileno, const uint32_t offset)
     uint32_t sig, size;
-    if (!datfarm.read(fileno, offset-8, sizeof(sig), &sig)) {
+    uint32_t fileno=FOFF[bk_no].fileno, offset=FOFF[bk_no].offset;
+    if (!datfarm->read(fileno, offset-8, sizeof(sig), &sig)) {
         cerr << "Can't read bk signature." << endl;
         return false;
     }
@@ -89,7 +63,7 @@ bool    load_bk(DATFARM_T &datfarm, const uint32_t fileno, const uint32_t offset
         cerr << "Block signature not found: " << hex << sig << endl;
         return false;
     }
-    if (!datfarm.read(fileno, offset-4, sizeof(size), &size)) {
+    if (!datfarm->read(fileno, offset-4, sizeof(size), &size)) {
         cerr << "Can't read bk size." << endl;
         return false;
     }
@@ -97,9 +71,7 @@ bool    load_bk(DATFARM_T &datfarm, const uint32_t fileno, const uint32_t offset
         cerr << "Block too big: " << size << endl;
         return false;
     }
-    //BUFFER.end = BUFFER.beg + size;
-    CUR_PTR.v_ptr = BUFFER.beg;
-    return datfarm.read(fileno, offset, size, BUFFER.beg);
+    return datfarm->read(fileno, offset, size, dst);
 }
 
 //fs::path sDatName = OPTS.bkdir + "/" + argv[file1idx];
